pyproject.toml
[project]
name = "fastapi-todo"
version = "0.1.0"
description = "Exercise app."
authors = [
    {name = "Błażej Jastrzębski",email = "jastrzebskib@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # BE framework, utils etc.
    "fastapi[standard] (>=0.117.1, <0.118.0)",
    "pydantic-settings (>=2.11.0, <2.12)",
    # ORM, db connections etc.
    "SQLAlchemy (>=2.0.44, <2.1)",
    "alembic (>=1.17.1, <1.18)",
    "psycopg (>=3.2.12, <3.13)",
    # Security
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
]

[dependency-groups]
test = [
    "pytest (>=8.4.2, <8.5)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
